fixed algorithm
fixed that for some reason it sent the same path array to all of the first nodes.
fixed that you can have a date pair with the last day of a not flaxable date and the next date.
fixed that if a location is shown more then once it will add the days in this location